api1/requirements.txt
fastapi==0.109.2
uvicorn==0.30.1
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.1
requests==2.31.0